+++ src/lib.rs
@@ -2,6 +2,7 @@ mod data;
 
 pub mod args;
 pub mod clap;
+pub mod config;
 pub mod filesystem;
 pub mod path;
 pub mod progress;