+++ src/path.rs
@@ -6,99 +6,157 @@ use ge_man_lib::tag::TagKind;
 #[cfg(test)]
 use mockall::automock;
 
+use crate::config;
+
 pub const STEAM_COMP_DIR: &str = "Steam/compatibilitytools.d";
 pub const LUTRIS_WINE_RUNNERS_DIR: &str = "lutris/runners/wine";
 
-const HOME: &str = "HOME";
-const XDG_DATA_HOME: &str = "XDG_DATA_HOME";
-const XDG_CONFIG_HOME: &str = "XDG_CONFIG_HOME";
-const STEAM_PATH_ENV: &str = "STEAM_PATH";
-
+pub const HOME: &str = "HOME";
+pub const XDG_DATA_HOME: &str = "XDG_DATA_HOME";
+pub const XDG_CONFIG_HOME: &str = "XDG_CONFIG_HOME";
+pub const STEAM_PATH_ENV: &str = "GE_MAN_STEAM_PATH";
 const APP_NAME: &str = "ge_man";
 
-pub fn xdg_data_home() -> Option<String> {
-    env::var(XDG_DATA_HOME).ok()
+pub fn xdg_data_home() -> Option<PathBuf> {
+    env::var(XDG_DATA_HOME).map(PathBuf::from).ok()
 }
 
-pub fn xdg_config_home() -> Option<String> {
-    env::var(XDG_CONFIG_HOME).ok()
+pub fn xdg_config_home() -> Option<PathBuf> {
+    env::var(XDG_CONFIG_HOME).map(PathBuf::from).ok()
 }
 
-pub fn steam_path() -> Option<String> {
-    env::var(STEAM_PATH_ENV).ok()
+pub fn steam_root() -> Option<PathBuf> {
+    env::var(STEAM_PATH_ENV)
+        .map(PathBuf::from)
+        .ok()
+        .or_else(|| config::GE_MAN_CONFIG.lock().unwrap().steam_root_path())
+}
+
+pub struct PathOverrides {
+    steam_root_cfg: Option<PathBuf>,
+    steam_root_env: Option<PathBuf>,
+    xdg_data_home: Option<PathBuf>,
+    xdg_config_home: Option<PathBuf>,
+}
+
+impl PathOverrides {
+    pub fn new(
+        steam_root_cfg: Option<PathBuf>,
+        steam_root_env: Option<PathBuf>,
+        xdg_data_home: Option<PathBuf>,
+        xdg_config_home: Option<PathBuf>,
+    ) -> Self {
+        Self {
+            steam_root_cfg,
+            steam_root_env,
+            xdg_data_home,
+            xdg_config_home,
+        }
+    }
+
+    pub fn steam_root(&self) -> Option<&PathBuf> {
+        self.steam_root_cfg.as_ref().or(self.steam_root_env.as_ref())
+    }
+
+    pub fn xdg_data_home(&self) -> Option<&PathBuf> {
+        self.xdg_data_home.as_ref()
+    }
+
+    pub fn xdg_config_home(&self) -> Option<&PathBuf> {
+        self.xdg_config_home.as_ref()
+    }
+}
+
+impl Default for PathOverrides {
+    fn default() -> Self {
+        let xdg_data = env::var(XDG_DATA_HOME).map(PathBuf::from).ok();
+        let xdg_config = env::var(XDG_CONFIG_HOME).map(PathBuf::from).ok();
+        let steam_root_env = env::var(STEAM_PATH_ENV).map(PathBuf::from).ok();
+        let steam_root_from_cfg = config::GE_MAN_CONFIG.lock().unwrap().steam_root_path();
+
+        PathOverrides::new(steam_root_from_cfg, steam_root_env, xdg_data, xdg_config)
+    }
 }
 
 #[cfg_attr(test, automock)]
 pub trait PathConfiguration {
-    fn xdg_data_dir(&self, xdg_data_home: Option<String>) -> PathBuf {
+    fn xdg_data_dir(&self, xdg_data_home: Option<PathBuf>) -> PathBuf {
         let data_dir = xdg_data_home
-            .or_else(|| env::var(HOME).ok().map(|home| format!("{}/.local/share", home)))
+            .or_else(|| {
+                env::var(HOME)
+                    .ok()
+                    .map(|home| PathBuf::from(format!("{}/.local/share", home)))
+            })
             .unwrap();
 
         PathBuf::from(data_dir)
     }
 
-    fn xdg_config_dir(&self, xdg_config_home: Option<String>) -> PathBuf {
-        let config_dir = xdg_config_home
-            .or_else(|| env::var(HOME).ok().map(|home| format!("{}/.config", home)))
-            .unwrap();
-
-        PathBuf::from(config_dir)
+    fn xdg_config_dir(&self, xdg_config_home: Option<PathBuf>) -> PathBuf {
+        xdg_config_home
+            .or_else(|| {
+                env::var(HOME)
+                    .ok()
+                    .map(|home| PathBuf::from(format!("{}/.config", home)))
+            })
+            .unwrap()
     }
 
-    fn steam(&self, steam_root_path_override: Option<String>) -> PathBuf {
+    fn steam(&self, steam_root_path_override: Option<PathBuf>) -> PathBuf {
         let steam_root_symlink = env::var(HOME)
             .ok()
             .map(|home| format!("{}/.steam/root", home))
             .map(PathBuf::from)
             .unwrap();
 
-        steam_root_path_override
-            .map(PathBuf::from)
-            .unwrap_or_else(|| steam_root_symlink)
+        steam_root_path_override.unwrap_or_else(|| steam_root_symlink)
     }
 
-    fn lutris_local(&self, xdg_data_home: Option<String>) -> PathBuf {
+    fn lutris_local(&self, xdg_data_home: Option<PathBuf>) -> PathBuf {
         self.xdg_data_dir(xdg_data_home).join("lutris")
     }
 
-    fn lutris_config(&self, xdg_config_home: Option<String>) -> PathBuf {
+    fn lutris_config(&self, xdg_config_home: Option<PathBuf>) -> PathBuf {
         self.xdg_config_dir(xdg_config_home).join("lutris")
     }
 
-    fn steam_config(&self, steam_root_path_override: Option<String>) -> PathBuf {
+    fn steam_config(&self, steam_root_path_override: Option<PathBuf>) -> PathBuf {
         self.steam(steam_root_path_override).join("config/config.vdf")
     }
 
-    fn steam_compatibility_tools_dir(&self, steam_root_path_override: Option<String>) -> PathBuf {
+    fn steam_compatibility_tools_dir(&self, steam_root_path_override: Option<PathBuf>) -> PathBuf {
         self.steam(steam_root_path_override).join("compatibilitytools.d")
     }
 
-    fn lutris_runners_config_dir(&self, xdg_config_home: Option<String>) -> PathBuf {
+    fn lutris_runners_config_dir(&self, xdg_config_home: Option<PathBuf>) -> PathBuf {
         self.lutris_config(xdg_config_home).join("runners")
     }
 
-    fn lutris_wine_runner_config(&self, xdg_config_home: Option<String>) -> PathBuf {
+    fn lutris_wine_runner_config(&self, xdg_config_home: Option<PathBuf>) -> PathBuf {
         self.lutris_runners_config_dir(xdg_config_home).join("wine.yml")
     }
 
-    fn lutris_runners_dir(&self, xdg_data_home: Option<String>) -> PathBuf {
+    fn lutris_runners_dir(&self, xdg_data_home: Option<PathBuf>) -> PathBuf {
         self.lutris_local(xdg_data_home).join("runners/wine")
     }
 
-    fn ge_man_data_dir(&self, xdg_data_home: Option<String>) -> PathBuf {
+    fn ge_man_data_dir(&self, xdg_data_home: Option<PathBuf>) -> PathBuf {
         self.xdg_data_dir(xdg_data_home).join(APP_NAME)
     }
 
-    fn ge_man_config_dir(&self, xdg_config_home: Option<String>) -> PathBuf {
+    fn ge_man_config_dir(&self, xdg_config_home: Option<PathBuf>) -> PathBuf {
         self.xdg_config_dir(xdg_config_home).join(APP_NAME)
     }
 
-    fn managed_versions_config(&self, xdg_data_home: Option<String>) -> PathBuf {
+    fn managed_versions_config(&self, xdg_data_home: Option<PathBuf>) -> PathBuf {
         self.ge_man_data_dir(xdg_data_home).join("managed_versions.json")
     }
 
-    fn app_config_backup_file(&self, xdg_config_home: Option<String>, kind: &TagKind) -> PathBuf {
+    fn ge_man_config(&self, xdg_config_home: Option<PathBuf>) -> PathBuf {
+        self.ge_man_config_dir(xdg_config_home).join("config.yml")
+    }
+
+    fn app_config_backup_file(&self, xdg_config_home: Option<PathBuf>, kind: &TagKind) -> PathBuf {
         let config_file = match kind {
             TagKind::Proton => "steam-config-backup.vdf",
             TagKind::Wine { .. } => "lutris-wine-runner-config-backup.yml",
@@ -106,7 +164,11 @@ pub trait PathConfiguration {
         self.ge_man_config_dir(xdg_config_home).join(config_file)
     }
 
-    fn create_ge_man_dirs(&self, xdg_config_home: Option<String>, xdg_data_home: Option<String>) -> anyhow::Result<()> {
+    fn create_ge_man_dirs(
+        &self,
+        xdg_config_home: Option<PathBuf>,
+        xdg_data_home: Option<PathBuf>,
+    ) -> anyhow::Result<()> {
         let ge_config_dir = self.ge_man_config_dir(xdg_config_home);
         let ge_data_dir = self.ge_man_data_dir(xdg_data_home);
 
@@ -124,9 +186,9 @@ pub trait PathConfiguration {
 
     fn create_app_dirs(
         &self,
-        xdg_config_home: Option<String>,
-        xdg_data_home: Option<String>,
-        steam_path: Option<String>,
+        xdg_config_home: Option<PathBuf>,
+        xdg_data_home: Option<PathBuf>,
+        steam_path: Option<PathBuf>,
     ) -> anyhow::Result<()> {
         let steam_compat_dir = self.steam_compatibility_tools_dir(steam_path);
         let lutris_runners_cfg_dir = self.lutris_runners_config_dir(xdg_config_home);
@@ -166,7 +228,7 @@ impl AppConfigPaths {
 impl<T: PathConfiguration> From<&T> for AppConfigPaths {
     fn from(path_cfg: &T) -> Self {
         AppConfigPaths::new(
-            path_cfg.steam_config(steam_path()),
+            path_cfg.steam_config(steam_root()),
             path_cfg.lutris_wine_runner_config(xdg_config_home()),
         )
     }
@@ -209,7 +271,7 @@ mod tests {
     #[test]
     fn get_xdg_data_path_with_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.xdg_data_dir(Some(String::from("/tmp/xdg-data")));
+        let path = path_cfg.xdg_data_dir(Some(PathBuf::from("/tmp/xdg-data")));
         assert_eq!(path, PathBuf::from("/tmp/xdg-data/"));
     }
 
@@ -225,7 +287,7 @@ mod tests {
     #[test]
     fn steam_path_with_steam_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.steam(Some(String::from("/tmp/steam")));
+        let path = path_cfg.steam(Some(PathBuf::from("/tmp/steam")));
 
         assert_eq!(path, PathBuf::from("/tmp/steam"));
     }
@@ -242,7 +304,7 @@ mod tests {
     #[test]
     fn lutris_local_path_with_xdg_data_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.lutris_local(Some(String::from("/tmp/xdg-data")));
+        let path = path_cfg.lutris_local(Some(PathBuf::from("/tmp/xdg-data")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-data/lutris"));
     }
@@ -259,7 +321,7 @@ mod tests {
     #[test]
     fn lutris_config_path_with_xdg_config_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.lutris_config(Some(String::from("/tmp/xdg-config")));
+        let path = path_cfg.lutris_config(Some(PathBuf::from("/tmp/xdg-config")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-config/lutris"));
     }
@@ -276,7 +338,7 @@ mod tests {
     #[test]
     fn steam_config_with_steam_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.steam_config(Some(String::from("/tmp/steam")));
+        let path = path_cfg.steam_config(Some(PathBuf::from("/tmp/steam")));
 
         assert_eq!(path, PathBuf::from("/tmp/steam/config/config.vdf"));
     }
@@ -293,7 +355,7 @@ mod tests {
     #[test]
     fn steam_compatibilitytools_with_steam_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.steam_compatibility_tools_dir(Some(String::from("/tmp/steam")));
+        let path = path_cfg.steam_compatibility_tools_dir(Some(PathBuf::from("/tmp/steam")));
 
         assert_eq!(path, PathBuf::from("/tmp/steam/compatibilitytools.d"));
     }
@@ -310,7 +372,7 @@ mod tests {
     #[test]
     fn lutris_wine_config_with_xdg_config_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.lutris_wine_runner_config(Some(String::from("/tmp/xdg-config")));
+        let path = path_cfg.lutris_wine_runner_config(Some(PathBuf::from("/tmp/xdg-config")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-config/lutris/runners/wine.yml"));
     }
@@ -327,7 +389,7 @@ mod tests {
     #[test]
     fn lutris_runners_dir_with_xdg_data_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.lutris_runners_dir(Some(String::from("/tmp/xdg-data")));
+        let path = path_cfg.lutris_runners_dir(Some(PathBuf::from("/tmp/xdg-data")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-data/lutris/runners/wine"));
     }
@@ -344,7 +406,7 @@ mod tests {
     #[test]
     fn ge_man_data_dir_with_xdg_data_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.ge_man_data_dir(Some(String::from("/tmp/xdg-data")));
+        let path = path_cfg.ge_man_data_dir(Some(PathBuf::from("/tmp/xdg-data")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-data/ge_man"));
     }
@@ -359,9 +421,26 @@ mod tests {
     }
 
     #[test]
+    fn ge_man_config_file_with_no_override() {
+        let path_cfg = PathConfig::default();
+        let path = path_cfg.ge_man_config(None);
+
+        assert!(path.to_string_lossy().contains("home"));
+        assert!(path.to_string_lossy().contains(".config/ge_man/config.yml"));
+    }
+
+    #[test]
+    fn ge_man_config_file_with_xdg_home_override() {
+        let path_cfg = PathConfig::default();
+        let path = path_cfg.ge_man_config(Some(PathBuf::from("/tmp/xdg-config")));
+
+        assert_eq!(path, PathBuf::from("/tmp/xdg-config/ge_man/config.yml"));
+    }
+
+    #[test]
     fn ge_man_config_dir_with_xdg_data_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.ge_man_config_dir(Some(String::from("/tmp/xdg-config")));
+        let path = path_cfg.ge_man_config_dir(Some(PathBuf::from("/tmp/xdg-config")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-config/ge_man"));
     }
@@ -380,7 +459,7 @@ mod tests {
     #[test]
     fn ge_man_managed_versions_with_xdg_config_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.managed_versions_config(Some(String::from("/tmp/xdg-config")));
+        let path = path_cfg.managed_versions_config(Some(PathBuf::from("/tmp/xdg-config")));
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-config/ge_man/managed_versions.json"));
     }
@@ -399,7 +478,7 @@ mod tests {
     #[test]
     fn ge_man_backup_file_for_steam_with_xdg_config_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.app_config_backup_file(Some(String::from("/tmp/xdg-config")), &TagKind::Proton);
+        let path = path_cfg.app_config_backup_file(Some(PathBuf::from("/tmp/xdg-config")), &TagKind::Proton);
 
         assert_eq!(path, PathBuf::from("/tmp/xdg-config/ge_man/steam-config-backup.vdf"));
     }
@@ -418,7 +497,7 @@ mod tests {
     #[test]
     fn ge_man_backup_file_for_lutris_with_xdg_config_override() {
         let path_cfg = PathConfig::default();
-        let path = path_cfg.app_config_backup_file(Some(String::from("/tmp/xdg-config")), &TagKind::wine());
+        let path = path_cfg.app_config_backup_file(Some(PathBuf::from("/tmp/xdg-config")), &TagKind::wine());
 
         assert_eq!(
             path,
@@ -433,12 +512,7 @@ mod tests {
         let data_dir = tmp_dir.join("local");
 
         let path_cfg = PathConfig::default();
-        path_cfg
-            .create_ge_man_dirs(
-                Some(config_dir.display().to_string()),
-                Some(data_dir.display().to_string()),
-            )
-            .unwrap();
+        path_cfg.create_ge_man_dirs(Some(config_dir), Some(data_dir)).unwrap();
 
         tmp_dir.child("config/ge_man").assert(predicates::path::exists());
         tmp_dir.child("local/ge_man").assert(predicates::path::exists());
@@ -455,11 +529,7 @@ mod tests {
 
         let path_cfg = PathConfig::default();
         path_cfg
-            .create_app_dirs(
-                Some(config_dir.display().to_string()),
-                Some(data_dir.display().to_string()),
-                Some(steam_dir.display().to_string()),
-            )
+            .create_app_dirs(Some(config_dir), Some(data_dir), Some(steam_dir))
             .unwrap();
 
         tmp_dir