+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+
+* Add the ability to set the Steam installation path via a config file in `$XDG_CONFIG_HOME/ge_man/config.json`.
+  Alternatively, the Steam path can also be set with the `GE_MAN_STEAM_PATH` variable.
+
 ## [0.1.2] - 2022-06-17
 
 ### Changed